# Trie node memory optimization