This PR reduces the size of a single node by `16 bytes`, from `80 bytes` to `64 bytes`. This impact the timing though. 
Replaces #4581

```bash
BenchmarkDotNet=v0.13.4, OS=Windows 11 (10.0.22621.1105)
12th Gen Intel Core i9-12900HK, 1 CPU, 20 logical and 14 physical cores
.NET SDK=7.0.100
  [Host]     : .NET 7.0.0 (7.0.22.51805), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.0 (7.0.22.51805), X64 RyuJIT AVX2
```

|        Method |     Mean |     Error |    StdDev |   Gen0 | Allocated |
|-------------- |---------:|----------:|----------:|-------:|----------:|
| PropertiesSet (before) | 7.509 ns | 0.0954 ns | 0.0846 ns | 0.0064 |      80 B |
| PropertiesSet (after) | 7.081 ns | 0.0677 ns | 0.0633 ns | 0.0051 |      64 B |

## Changes:

- collapsing numeric values into a single field reduces object size from `80B` to `72B` 
- collapsing `IsBoundaryProofNode` moves it from `72B` to `64B`
- fixing `GetMemorySize` which previously was sharing a wrong value (no `LastSeen` reported at all!)
- `LastSeen` made non-nullable which makes the bit operations much easier

## Types of changes

What types of changes does your code introduce?
_Put an `x` in the boxes that apply_

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation Update
- [ ] Code style update (formatting, renaming)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Other (please describe): 

## Testing
**Requires testing**

- [ ] Yes
- [x] No

**In case you checked yes, did you write tests??**

- [x] Yes
- [ ] No