@Scooletz 2023-02-09
You are correct @tkstanczak There's some complexity added to the `TrieNode` which results in drop of perf. The memory savings, the question is whether it's worth to work on it and save 16b per node. Let me take a deeper look.

BTW added an example of ASM to the description so that it shows the difference.