@benaadams 2023-02-09
```asm
; after
TrieNode.set_NodeType(NodeType)
    L0000: push esi
    L0001: lea eax, [ecx+4]
    L0004: mov esi, [eax]
    L0006: mov eax, [eax+4]
    L0009: and esi, 0xfffffffc
    L000c: and eax, 0xffffffff   ; <--- ? Could be (int) cast
    L000f: movzx edx, dl
    L0012: or edx, esi
    L0014: or eax, 0             ; <--- ? Does nothing?
    L0017: add ecx, 4
    L001a: mov [ecx], edx
    L001c: mov [ecx+4], eax
    L001f: pop esi
    L0020: ret
```
Two of the asm lines are a bit weird