@Scooletz 2023-02-09
@benaadams I won't be chasing these as after working on `LastSeen` and running benchmarks on a non-busy machine, I get similar speed with initial bytes saving. 

@LukaszRozmej Could you re-review? The `LastSeen` has been changed.

@tkstanczak Fixed the timing.