@benaadams 2023-02-09
> @benaadams I won't be chasing these as after working on `LastSeen` and running benchmarks on a non-busy machine, I get similar speed with initial bytes saving.

I see it as 
```asm
; Assembly listing for method Nethermind.Trie.TrieNode:set_NodeType(ubyte):this
; Emitting BLENDED_CODE for X64 CPU with AVX - Windows
; optimized code
; rsp based frame
; partially interruptible
; No PGO data

G_M000_IG01:                ;; offset=0000H
 
G_M000_IG02:                ;; offset=0000H
       488B4130             mov      rax, qword ptr [rcx+30H]
       4883E0FC             and      rax, -4
       0FB6D2               movzx    rdx, dl
       480BC2               or       rax, rdx
       48894130             mov      qword ptr [rcx+30H], rax
 
G_M000_IG03:                ;; offset=0012H
       C3                   ret      
 
; Total bytes of code 19
```

Do you use the `Disasmo` extension by @EgorBo? Recommend giving it a try if not, right click method show asm https://marketplace.visualstudio.com/items?itemName=EgorBogatov.Disasmo